+++ xenon-common/src/main/java/com/vmware/xenon/common/ServiceDocumentDescription.java
@@ -19,7 +19,6 @@ import java.lang.reflect.Modifier;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
 import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Date;
@@ -30,6 +29,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 import com.vmware.xenon.common.RequestRouter.Route;
 import com.vmware.xenon.common.Service.Action;
@@ -141,6 +141,10 @@ public class ServiceDocumentDescription {
 
     public static class PropertyDescription {
         public ServiceDocumentDescription.TypeName typeName;
+        /**
+         * Set only for PODO-typed fields.
+         */
+        public String kind;
         public Object exampleValue;
         transient Field accessor;
 
@@ -157,6 +161,11 @@ public class ServiceDocumentDescription {
          * Description of element type if this property is of the ARRAY or COLLECTION type.
          */
         public PropertyDescription elementDescription;
+
+        /**
+         * Set only for enums, the set of possible values.
+         */
+        public String[] enumValues;
 
         public PropertyDescription() {
             this.indexingOptions = EnumSet.noneOf(PropertyIndexingOption.class);
@@ -221,6 +230,10 @@ public class ServiceDocumentDescription {
 
             desc.propertyDescriptions = root.fieldDescriptions;
             return desc;
+        }
+
+        public PropertyDescription buildPodoPropertyDescription(Class<?> type) {
+            return buildPodoPropertyDescription(type, new HashSet<>(), 0);
         }
 
         public ServiceDocumentDescription buildDescription(
@@ -288,6 +301,10 @@ public class ServiceDocumentDescription {
 
                 fd.accessor = f;
                 pd.fieldDescriptions.put(f.getName(), fd);
+
+                if (fd.typeName == TypeName.PODO) {
+                    fd.kind = Utils.buildKind(f.getType());
+                }
             }
 
             visited.remove(typeName);
@@ -328,6 +345,10 @@ public class ServiceDocumentDescription {
             } else if (Float.class.equals(clazz) || float.class.equals(clazz)) {
                 pd.typeName = TypeName.DOUBLE;
                 pd.exampleValue = 0.0F;
+            } else if (Number.class.isAssignableFrom(clazz)) {
+                // coerce all Numbers to double, may lose precision with BigDecimal and BigInteger
+                pd.typeName = TypeName.DOUBLE;
+                pd.exampleValue = 0.0F;
             } else if (String.class.equals(clazz) || char.class.equals(clazz)) {
                 pd.typeName = TypeName.STRING;
                 pd.exampleValue = "example string";
@@ -335,10 +356,7 @@ public class ServiceDocumentDescription {
                 pd.exampleValue = new Date();
                 pd.typeName = TypeName.DATE;
             } else if (URI.class.equals(clazz)) {
-                try {
-                    pd.exampleValue = new URI("http://localhost:1234/some/service");
-                } catch (URISyntaxException ignored) {
-                }
+                pd.exampleValue = URI.create("http://localhost:1234/some/service");
                 pd.typeName = TypeName.URI;
             } else {
                 isSimpleType = false;
@@ -422,6 +440,11 @@ public class ServiceDocumentDescription {
                     pd.elementDescription = fd;
                 } else if (Enum.class.isAssignableFrom(clazz)) {
                     pd.typeName = TypeName.ENUM;
+                    pd.enumValues = Arrays
+                            .stream(clazz.getEnumConstants())
+                            .map( o -> ((Enum)o).name())
+                            .collect(Collectors.toList())
+                            .toArray(new String[0]);
                 } else if (clazz.isArray()) {
                     pd.typeName = TypeName.ARRAY;
 
@@ -461,6 +484,7 @@ public class ServiceDocumentDescription {
                     pd.elementDescription = fd;
                 } else {
                     pd.typeName = TypeName.PODO;
+                    pd.kind = Utils.buildKind(clazz);
                     if (depth > 0) {
                         // Force indexing of all nested complex PODO fields. If the service author
                         // instructed expand at the root level, we will index and expand everything