+++ xenon-samples/src/main/java/com/vmware/xenon/samples/SampleHost.java
@@ -15,12 +15,19 @@ package com.vmware.xenon.samples;
 
 import java.util.logging.Level;
 
+
+import io.swagger.models.Contact;
+import io.swagger.models.Info;
+import io.swagger.models.License;
+
 import com.vmware.xenon.common.ServiceHost;
 import com.vmware.xenon.services.common.RootNamespaceService;
+import com.vmware.xenon.services.common.ServiceUriPaths;
 import com.vmware.xenon.services.samples.SampleFactoryServiceWithCustomUi;
 import com.vmware.xenon.services.samples.SamplePreviousEchoFactoryService;
 import com.vmware.xenon.services.samples.SampleServiceWithSharedCustomUi;
 import com.vmware.xenon.services.samples.SampleSimpleEchoFactoryService;
+import com.vmware.xenon.swagger.SwaggerDescriptorService;
 import com.vmware.xenon.ui.UiService;
 
 /**
@@ -67,6 +74,27 @@ public class SampleHost extends ServiceHost {
 
         // Start UI service
         super.startService(new UiService());
+
+        // Serve Swagger 2.0 compatible API description
+        SwaggerDescriptorService swagger = new SwaggerDescriptorService();
+
+        // exclude some core services
+        swagger.setExcludedPrefixes(
+                "/core/transactions",
+                "/core/node-groups");
+
+        // Provide API metainfo
+        Info apiInfo = new Info();
+        apiInfo.setVersion("1.0.0");
+        apiInfo.setTitle("Xenon SampleHost");
+        apiInfo.setLicense(new License().name("Apache 2.0").url("https://github.com/vmware/xenon/blob/master/LICENSE"));
+        apiInfo.setContact(new Contact().url("https://github.com/vmware/xenon"));
+        swagger.setInfo(apiInfo);
+
+        // Serve swagger on default uri
+        SwaggerDescriptorService.startService(this, swagger);
+        System.out.println("Checkout swaggerUI: " + this.getPublicUri() + ServiceUriPaths.SWAGGER + "/ui");
+
         return this;
     }
 }