+++ xenon-swagger/src/test/java/com/vmware/xenon/swagger/CarService.java
@@ -1,0 +1,59 @@
+/*
+ * Copyright (c) 2015 VMware, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License.  You may obtain a copy of
+ * the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed
+ * under the License is distributed on an "AS IS" BASIS, without warranties or
+ * conditions of any kind, EITHER EXPRESS OR IMPLIED.  See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package com.vmware.xenon.swagger;
+
+import java.net.URI;
+
+import com.vmware.xenon.common.FactoryService;
+import com.vmware.xenon.common.Operation;
+import com.vmware.xenon.common.Service;
+import com.vmware.xenon.common.ServiceDocument;
+import com.vmware.xenon.common.StatefulService;
+
+/**
+ */
+public class CarService extends StatefulService {
+    public static final String FACTORY_LINK = "/cars";
+
+    public CarService() {
+        super(Car.class);
+    }
+
+    public static Service createFactory() {
+        return FactoryService.create(CarService.class, Car.class);
+    }
+
+    public void handlePut(Operation put) {
+        this.setState(put, put.getBody(Car.class));
+        put.complete();
+    }
+
+    public enum FuelType {
+        GASOLINE, DIESEL, ELECTRICITY
+    }
+
+    public static class EngineInfo {
+        public FuelType fuel;
+        public double power;
+    }
+
+    public static class Car extends ServiceDocument {
+        public URI manufacturerHomePage;
+        public double length;
+        public double weight;
+        public String make;
+        public String licensePlate;
+        public EngineInfo engineInfo;
+    }
+}